-Added tests -Removed support for 3.5 (since aiohttp doesn't support it)
